Tried switching to pool
There doesn't seem to be a good way to make this work either because
I can't pickle the klampt rigid objects that need to be passed
into the coverage computation. If I want to introduce parallelism,
it seems that it's going to need to be in c/c++ because manually
creating Python processes is too heavy and requires serializing
different objects to pass between them (really, this is a job for
threads, but the GIL prevents that from really giving any speedup).