Debugged Heightmaps
Resolved some bugs in computing the sample points for the heightmaps.
Probably need to pass things like the extent of the map up to python as
well - should this be done in a separate function?

Restrict valid sample points to triangles for which a given plane is
responsible - otherwise you can get points in the heightmap that are
not well approximated by the plane.